Add file type check.
Former-commit-id: 9171f36fc5452d350331e9e3bb7048ac31f249aa

+++ backblaze/src/main/java/ch/cyberduck/core/b2/B2DeleteFeature.java
@@ -50,7 +50,7 @@ public class B2DeleteFeature implements Delete {
                         session.getClient().deleteFileVersion(String.format("%s/.bzEmpty", containerService.getKey(file)),
                                 new B2FileidProvider(session).getFileid(file));
                     }
-                    else {
+                    else if(file.isFile()) {
                         session.getClient().deleteFileVersion(containerService.getKey(file),
                                 new B2FileidProvider(session).getFileid(file));
                     }
